requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.95.0
h11==0.14.0
hazelcast-python-client==5.1
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
pydantic==1.10.4
python-consul==1.1.0
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
